controller_manager: resolve implicit cast errors

+++ controller_manager/src/controller_manager.cpp
@@ -245,7 +245,8 @@ ControllerManager::ControllerManager(
   bool activate_all_hw_components, const std::string & manager_node_name,
   const std::string & node_namespace, const rclcpp::NodeOptions & options)
 : rclcpp::Node(manager_node_name, node_namespace, options),
-  update_rate_(get_parameter_or<int>("update_rate", 100)),
+  update_rate_(static_cast<unsigned int>(
+    get_parameter_or<int>("update_rate", 100))),
   resource_manager_(std::make_unique<hardware_interface::ResourceManager>(
     urdf, this->get_node_clock_interface(), this->get_node_logging_interface(),
     activate_all_hw_components, update_rate_)),
